refactor: simplify test BindingsInsideProvider
