+++ tests/Feature/Framework/BindingsInsideProvider/Module/Application/GreeterService.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Feature\Framework\BindingsInsideProvider\Module\Domain;
+namespace GacelaTest\Feature\Framework\BindingsInsideProvider\Module\Application;
+
+use GacelaTest\Feature\Framework\BindingsInsideProvider\Module\Domain\GreeterGeneratorInterface;
 
 final class GreeterService
 {
