+++ tests/Feature/Framework/BindingsInsideProvider/Module/Domain/Greeter/Team.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace GacelaTest\Feature\Framework\BindingsInsideProvider\Module\Domain\Greeter;
 
-final class CustomNameGenerator
+final class Team
 {
     public function getNames(): string
     {
