+++ tests/Feature/Framework/BindingsInsideProvider/Module/Factory.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace GacelaTest\Feature\Framework\BindingsInsideProvider\Module;
 
 use Gacela\Framework\AbstractFactory;
+use GacelaTest\Feature\Framework\BindingsInsideProvider\Module\Application\GreeterService;
 use GacelaTest\Feature\Framework\BindingsInsideProvider\Module\Domain\GreeterGeneratorInterface;
-use GacelaTest\Feature\Framework\BindingsInsideProvider\Module\Domain\GreeterService;
 
 final class Factory extends AbstractFactory
 {
